# docs/swagger.yaml
definitions:
  model.Clinic:
    properties:
      clinic_id:
        example: 1
        maximum: 11
        type: integer
      deleted:
        example: 0
        maximum: 1
        type: integer
      end_at:
        example: "2099-12-31"
        type: string
      name:
        example: 診所
        maxLength: 128
        type: string
      predict:
        items:
          $ref: '#/definitions/model.Predict'
        type: array
      quota_per_month:
        example: 200
        maximum: 100000
        type: integer
      start_at:
        example: "2022-10-10"
        type: string
      token:
        example: token1231324568913
        maxLength: 1024
        type: string
    required:
    - clinic_id
    - end_at
    - name
    - quota_per_month
    - start_at
    - token
    type: object
  model.ClinicCreateForm:
    properties:
      end_at:
        example: "2099-12-31"
        type: string
      name:
        example: 診所
        maxLength: 128
        type: string
      quota_per_month:
        example: 200
        maximum: 100000
        type: integer
      start_at:
        example: "2022-10-10"
        type: string
    required:
    - end_at
    - name
    - quota_per_month
    - start_at
    type: object
  model.ClinicTokenForm:
    properties:
      clinic_id:
        example: 1
        maximum: 11
        type: integer
    required:
    - clinic_id
    type: object
  model.ClinicUpdateForm:
    properties:
      clinic_id:
        example: 1
        maximum: 11
        type: integer
      end_at:
        example: "2099-12-31"
        type: string
      name:
        example: 診所
        maxLength: 128
        type: string
      quota_per_month:
        example: 200
        maximum: 100000
        type: integer
      start_at:
        example: "2022-10-10"
        type: string
    required:
    - clinic_id
    - end_at
    - name
    - quota_per_month
    - start_at
    type: object
  model.Predict:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      dir:
        type: string
      filename:
        type: string
      predict:
        description: JSON要顯示，但是不要存入DB
        items:
          $ref: '#/definitions/model.PredictDetail'
        type: array
      predict_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.PredictDetail:
    properties:
      data:
        properties:
          bridge:
            items:
              type: integer
            type: array
          caries:
            items:
              type: integer
            type: array
          crown:
            items:
              type: integer
            type: array
          embedded:
            items:
              type: integer
            type: array
          endo:
            items:
              type: integer
            type: array
          filling:
            items:
              type: integer
            type: array
          imp:
            items:
              type: integer
            type: array
          impacted:
            items:
              type: integer
            type: array
          post:
            items:
              type: integer
            type: array
          r_r:
            items:
              type: integer
            type: array
        type: object
      filename:
        type: string
      text:
        type: string
    type: object
  model.User:
    properties:
      account:
        example: user
        maxLength: 64
        type: string
      name:
        example: User
        maxLength: 64
        type: string
      roles:
        example:
        - admin
        items:
          type: string
        maxItems: 1024
        type: array
      status:
        example: 1
        maximum: 11
        type: integer
      user_id:
        example: 1
        maximum: 11
        type: integer
    required:
    - account
    - name
    - roles
    - user_id
    type: object
  model.UserCreateForm:
    properties:
      account:
        example: user
        maxLength: 64
        type: string
      name:
        example: User
        maxLength: 64
        type: string
      passconf:
        example: password
        type: string
      password:
        description: 只能創建和更新
        example: password
        maxLength: 1024
        type: string
      roles:
        example:
        - admin
        items:
          type: string
        maxItems: 1024
        type: array
      status:
        example: 1
        maximum: 11
        type: integer
    required:
    - account
    - name
    - passconf
    - password
    - roles
    type: object
  model.UserLoginForm:
    properties:
      account:
        example: user
        maxLength: 64
        type: string
      password:
        example: password
        maxLength: 1024
        type: string
    required:
    - account
    - password
    type: object
  model.UserPasswordForm:
    properties:
      oldpass:
        description: 只能創建和更新
        example: old_password
        maxLength: 1024
        type: string
      passconf:
        description: 不對DB做任何事
        example: password
        type: string
      password:
        description: 只能創建和更新
        example: password
        maxLength: 1024
        type: string
      user_id:
        example: 1
        maximum: 11
        type: integer
    required:
    - passconf
    - password
    - user_id
    type: object
  model.UserToken:
    properties:
      token:
        type: string
    type: object
  model.UserUpdateAccountForm:
    properties:
      account:
        example: user
        maxLength: 64
        type: string
      roles:
        example:
        - admin
        items:
          type: string
        maxItems: 1024
        type: array
      status:
        example: 1
        maximum: 11
        type: integer
      user_id:
        example: 1
        maximum: 11
        type: integer
    required:
    - account
    - roles
    - user_id
    type: object
  model.UserUpdateForm:
    properties:
      name:
        example: User
        maxLength: 64
        type: string
      user_id:
        example: 1
        maximum: 11
        type: integer
    required:
    - name
    - user_id
    type: object
  utils.IH200:
    properties:
      data: {}
      msg:
        default: 成功
        type: string
      status:
        default: true
        type: boolean
    type: object
  utils.IH500:
    properties:
      msg:
        default: desc...
        type: string
      status:
        default: false
        type: boolean
    type: object
  utils.MyClaims:
    properties:
      aud:
        type: string
      exp:
        type: integer
      iat:
        type: integer
      iss:
        type: string
      jti:
        type: string
      name:
        type: string
      nbf:
        type: integer
      roles:
        items:
          type: string
        type: array
      sub:
        type: string
      user_id:
        type: integer
    type: object
host: localhost
info:
  contact: {}
  description: Swagger API.
  title: Gin Go Pano
  version: "1.0"
paths:
  /api/clinic:
    delete:
      consumes:
      - application/json
      operationId: "15"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ClinicTokenForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 刪除診所
      tags:
      - Clinic
    get:
      consumes:
      - application/json
      operationId: "7"
      produces:
      - application/json
      responses:
        "200":
          description: 診所
          schema:
            allOf:
            - $ref: '#/definitions/utils.IH200'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Clinic'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 獲取所有診所
      tags:
      - Clinic
    post:
      consumes:
      - application/json
      operationId: "8"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ClinicCreateForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 新增診所
      tags:
      - Clinic
    put:
      consumes:
      - application/json
      operationId: "9"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ClinicUpdateForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新診所資訊
      tags:
      - Clinic
  /api/clinic/token:
    patch:
      consumes:
      - application/json
      operationId: "10"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ClinicTokenForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新診所Token
      tags:
      - Clinic
  /api/predict:
    post:
      operationId: "1"
      parameters:
      - description: 請使用診所ID
        in: formData
        name: clinic_id
        required: true
        type: integer
      - description: 請選擇牙齒的X光圖
        in: formData
        name: nhicode
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 辨識資料
          schema:
            allOf:
            - $ref: '#/definitions/utils.IH200'
            - properties:
                data:
                  $ref: '#/definitions/model.Predict'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      summary: GRPC上傳圖片和AI辨識
      tags:
      - Predict
  /api/user:
    delete:
      consumes:
      - application/json
      operationId: "14"
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 刪除使用者
      tags:
      - User
    get:
      consumes:
      - application/json
      operationId: "2"
      produces:
      - application/json
      responses:
        "200":
          description: 使用者
          schema:
            allOf:
            - $ref: '#/definitions/utils.IH200'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 獲取所有使用者
      tags:
      - User
    patch:
      consumes:
      - application/json
      operationId: "4"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新使用者資訊
      tags:
      - User
    post:
      consumes:
      - application/json
      operationId: "3"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserCreateForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 新增使用者
      tags:
      - User
    put:
      consumes:
      - application/json
      operationId: "15"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateAccountForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新使用者帳號資訊
      tags:
      - User
  /api/user/info:
    get:
      consumes:
      - application/json
      operationId: "13"
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.IH200'
            - properties:
                data:
                  $ref: '#/definitions/utils.MyClaims'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 使用者JWT資訊
      tags:
      - User
  /api/user/login:
    post:
      consumes:
      - application/json
      operationId: "11"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.IH200'
            - properties:
                data:
                  $ref: '#/definitions/model.UserToken'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      summary: 使用者登入
      tags:
      - User
  /api/user/password:
    patch:
      consumes:
      - application/json
      operationId: "5"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新其他使用者密碼
      tags:
      - User
  /api/user/password/self:
    patch:
      consumes:
      - application/json
      operationId: "12"
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UserPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.IH200'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.IH500'
      security:
      - BearerAuth: []
      summary: 更新自己的密碼
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: 後台介面 - 使用者Api
  name: User
- description: 後台介面 - 診所Api
  name: Clinic
- description: 診所端 - 辨識Api
  name: Predict